@@ -1,8 +1,8 @@
-package com.example.tyrone.lowsignalwarning;
+package com.moraga.tyrone.lowsignalwarning;
 
 import android.content.Context;
-import android.support.test.InstrumentationRegistry;
-import android.support.test.runner.AndroidJUnit4;
+import androidx.test.InstrumentationRegistry;
+import androidx.test.ext.junit.runners.AndroidJUnit4;
 
 import org.junit.Test;
 import org.junit.runner.RunWith;
@@ -17,10 +17,11 @@ import static org.junit.Assert.*;
 @RunWith(AndroidJUnit4.class)
 public class ExampleInstrumentedTest {
     @Test
+
     public void useAppContext() throws Exception {
         // Context of the app under test.
         Context appContext = InstrumentationRegistry.getTargetContext();
 
-        assertEquals("com.example.tyrone.lowsignalwarning", appContext.getPackageName());
+        assertEquals("com.moraga.tyrone.lowsignalwarning", appContext.getPackageName());
     }
 }
