@@ -1,4 +1,4 @@
-package com.example.tyrone.lowsignalwarning;
+package com.moraga.tyrone.lowsignalwarning;
 
 
 import android.app.Service;
@@ -18,15 +18,15 @@ import android.util.Log;
  * Created by Tyrone on 8/1/19.
  */
 
-public class CellServiceListener extends Service{
-    public CellServiceListener(){
+public class ServiceListener extends Service{
+    public ServiceListener(){
 
     }
     private TelephonyManager telephonyManager;
     private PhoneStateListener listener;
     private int delayTime = 5000;
     private boolean goodService = true;
-    private String TAG = "CellServiceListener";
+    private String TAG = "ServiceListener";
 
     public IBinder onBind(Intent arg0){
         return null;
@@ -41,31 +41,34 @@ public class CellServiceListener extends Service{
 
     @Override
     public int onStartCommand(Intent intent, int flags, int startId){
-        final byte notifyBy = intent.getByteExtra(getString(R.string.signal_key), (byte)1);
+        final byte notifyBy = intent.getByteExtra(getString(R.string.signal_key), (byte)0);
+        final int serviceThreshold = intent.getIntExtra("Service Threshold", CellSignalStrength.SIGNAL_STRENGTH_POOR);
+        Log.v(TAG, "service threshold = " + serviceThreshold);
 
         listener = new PhoneStateListener(){
             @Override
             public void onSignalStrengthsChanged(SignalStrength signalStrength){
                 int signalLevel = signalStrength.getLevel();
 
-                if(goodService && signalLevel == CellSignalStrength.SIGNAL_STRENGTH_POOR){
+                if(goodService && signalLevel <= serviceThreshold){
                     Utils.showToast("Bad service detected", getApplicationContext());
                     Log.v(TAG, "Bad service detected");
                     goodService = false;
-                    serviceChange(notifyBy, Utils.badSignalPattern);
+                    serviceChange(notifyBy, Utils.badSignalPattern, goodService);
                     try{
                         Thread.sleep(delayTime);
+                        Log.v(TAG, "Mandated sleep");
                     }
                     catch(InterruptedException e){
                         Log.v(TAG, "Thread.sleep interrupted");
                     }
                 }
 
-                if(!goodService && signalLevel > CellSignalStrength.SIGNAL_STRENGTH_POOR){
+                if(!goodService && signalLevel > serviceThreshold){
                     goodService = true;
                     Utils.showToast("Good service detected", getApplicationContext());
                     Log.v(TAG, "Good service detected");
-                    serviceChange(notifyBy, Utils.goodSignalPattern);
+                    serviceChange(notifyBy, Utils.goodSignalPattern, goodService);
                 }
             }
         };
@@ -74,8 +77,7 @@ public class CellServiceListener extends Service{
         return START_STICKY;
     }
 
-    private void serviceChange(byte notifyBy, long pattern []){
-        Log.v(TAG, "notifyBy = " + notifyBy);
+    private void serviceChange(byte notifyBy, long pattern [], boolean service){
         switch(notifyBy){
             case 1:
                 serviceVibrate(pattern);
@@ -85,19 +87,29 @@ public class CellServiceListener extends Service{
             case 2:
                 serviceFlash(pattern);
                 break;
+            case 0:
+                Utils.showToast( "Good service?: " + service, this.getApplicationContext());
+        }
+        try{
+            Thread.sleep(delayTime);
+            Log.v(TAG, "service change sleep");
+        } catch (InterruptedException e){
+            e.printStackTrace();
         }
     }
 
     private void serviceFlash(long pattern[])
     {
-        Log.v(TAG, "service flash called");
         CameraManager cameraManager = (CameraManager) getSystemService(Context.CAMERA_SERVICE);
         for (int i = 0; i < pattern.length; i++) {
             if (i % 2 == 0) {
                 try {
                     String cameraId = cameraManager.getCameraIdList()[0];
                     cameraManager.setTorchMode(cameraId, true);
-                } catch (CameraAccessException e) {}
+                } catch (CameraAccessException e) {
+                    //if previous checks are broken
+                    Log.v(TAG, "Can't Flash");
+                }
             }
             else {
                 try {
@@ -118,9 +130,9 @@ public class CellServiceListener extends Service{
         Vibrator v = (Vibrator) getSystemService(Context.VIBRATOR_SERVICE);
 
         if (v.hasVibrator()) {
-            Log.v("Can Vibrate", "YES");
+            Log.v(TAG,"Can Vibrate");
         } else {
-            Log.v("Can Vibrate", "NO");
+            Log.v(TAG,"Can't Vibrate");
         }
         v.vibrate(pattern, -1);
     }
